ColorPickerPreference: Introduce set default button when attribute is set
Set a default value using the "defaultColorValue" settings namespace
attribute. The general idea is often we want set a value of say "-1"
and let framework handle it. One such scenario would be if you are
changing the color of a ImageView using ColorFilter. When framework
receives a "-1" value, framework sets the ImageView ColorFilter to null,
restoring the ImageView to original state.

Screenshot:

https://teameos.slack.com/files/bigrushdog/F09T5S155/screenshot_2015-08-31-04-02-54.png

Change-Id: I4ad63576270de344a9e7430d9e63a248d37afd9f

Change-Id: I9fa21eb8c7b12372e4a5cfd08d92d2ea45da0939
Signed-off-by: spezi77 <spezi7713@gmx.net>
Signed-off-by: NurKeinNeid <mralexman3000@gmail.com>